pyproject.toml
[project]
name = "pngsvg"
version = "0.1.0"
description = ""
authors = [
    {name = "dimash",email = "arystambekdimash005@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "uvicorn (>=0.40.0,<0.41.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-telegram-bot (>=22.5,<23.0)",
    "rembg[cpu] (>=2.0.72,<3.0.0)",
    "google-genai (>=1.59.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "openai (>=2.15.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
